scripts/requirements.txt
# Python dependencies for VideoTile scraper
# Install: pip install -r requirements.txt

requests>=2.31.0
beautifulsoup4>=4.12.0
supabase>=2.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
